+++ cloud_recording_test.go
@@ -29,3 +29,24 @@ func TestListAllRecordings(t *testing.T) {
 		t.Fatalf("got error listing recordings: %+v\n", err)
 	}
 }
+
+func TestGetMeetingRecordingsNoRecording(t *testing.T) {
+	var (
+		apiKey    = os.Getenv("ZOOM_API_KEY")
+		apiSecret = os.Getenv("ZOOM_API_SECRET")
+	)
+
+	APIKey = apiKey
+	APISecret = apiSecret
+
+	_, err := GetMeetingRecordings(GetMeetingRecordingsOptions{
+		MeetingID: "FooBar",
+	})
+
+	expected := `Zoom API error 3301: "There is no recording for this meeting"`
+	actual := err.Error()
+
+	if actual != expected {
+		t.Errorf(`Expected error "%s". Got "%s".`, expected, actual)
+	}
+}
